Django/3.2-crud/stocks_products/requirements.txt
django~=4.0.2
djangorestframework~=3.13.1
psycopg2-binary

django-filter~=21.1
django-debug-toolbar~=3.2.4